Move spinner-button to component library
